fix(network): makes device flag optional
The device_name_to_guid decorator only worked if device_name is present.
This assumption is correct for all the device sub-commands where
device_name is the argument.

In Network though, device_name is only required for device networks. So,
this commit adds the check to name device_name optional in the
decorator.