feat(project): adds support for specifying organization in create project command (#54)
The `project create` command now accepts `--organization` flag. It accepts both name and organization GUID.  This allows user to create a project in other organizations user is part of but is not currently operating under.

If this flag is not provided, then organization of the currently selected project will be used. User can check the current project through multiple ways:
1. `project list` command highlights the current project in green color.
2. `shell` command uses project name as the shell prompt.
3. The config file saves the `project_id` variable as well.

Resolves #48