feat(utils): adds decorator for yaspin spinner
The click spinner doesn't look as good and doesn't offer any additional
features out of the box. yaspin is a better package and similar spinners
are used across command-line based products these days.

This commit implements a decorator that can be used with commands where
showing a spinner makes sense. The decorator simply wraps a function
with the yaspin context-manager and the wrapped function can use the
spinner instance without having to worry about starting and stopping it.

Closes #113