fix(deployment): adds deployment as component alias to avoid conflicts
The packages generated using the Apply-method have the Component set to
'default'.

In the Cloud, if two deployments using Routed Network with
InboundScopedTargeted set to True are deployed with such a Package where
one Deployment depends on the other, then both Deployments will use the
Component-name ('default' here) as the Alias. This will cause Conflict
in the aliases. This will prevent the Deployment creation with
BadRequest.

On the Device, if two deployments using Routed Network with
InboundScopedTargeted set to True are deployed on the same device, then
both Deployments will use the Device-name as the Alias. This will again
cause the Conflict in the aliases.

This commit explicitly sets Deployment name as the Alias. For
Deployments not using InboundScopedTargeted, this will be a Noop, but
for others it will ensure that Alias is unique as we have
UniqueConstraint on the Deployment name.

Wrike: https://www.wrike.com/open.htm?id=1006380505
Github: https://github.com/rapyuta-robotics/rapyuta-io-cli/issues/68