fix(apply): fixes the schema not found errors
The recent change in JSON Schema validation made it such that the
Schemas are loaded at runtime. But, the Schemas were not being shipped
in the Python wheel. This was leading to schema not found error upon
running the apply command.

This commit adds schemas as part of the shipped Wheel to fix the issue.