fix(apply): reads files within nested directories
The apply command does not read nested directories well and hence, it is
not possible to break down manifests into sub-directories and apply them
by just passing the root directory as the input along with values and
secrets. The traversal does happen while the files are processed,
however the final list of files ends up with directories which cannot be
processed during the apply phase.

This commit filters out all the files and returns them omitting any
directories.