fix(project): fixes project update with vpn state
Recently, the project update API was updated to reflect the state as
sent in the request payload. The vpn command in project had the older
implementation and we forgot to update it when the project update API
was changed. This issue can lead to serious loss of project info if not
fixed in time. This commit corrects that by first fetching the project
and updating it with the new values before passing it as the payload for
update.