feat(device)!: accepts routes to advertise when enabling VPN
This commit changes the --advertise-routes option to accept one more
routes to advertise on a device. The idea behind doing this is that a device
may or may not advertise all the routes configured with a project. You
can specify only a set of routes that you want a device to advertise.