fix(apply): adds deploymentId as an option in GUID_KEYS
Deployment depends in manifest was not working when the deployment guid
was used instead of the deployment name in nameOrGUID. This was
happening because we were not accounting for 'deploymentId' for
dependent deployments and hence, apply was failing with the error that
there is a missing resource in the manifest.

This commit ensures that we read the right key and populate the
dependent deployment correctly in the apply hierarchy.

Wrike Ticket: https://www.wrike.com/open.htm?id=1249688483