fix(jsonschema): removes default device deployment restart policy
This default in the deployments was conflicting with the restart policy
set in the package. For example, if you set restart policy = never in
the package and you do not set the policy in the deployment manifest, it
still overrides the package value and sets the default.

Wrike Ticket:  https://www.wrike.com/open.htm?id=1205809047