fix(parameter): throws error when param tree not found during apply
Applying config parameter trees that do not exist did not throw an
error. This gave an impression to the end user that the operation was
successful. While this may make the command seem idempotent, it may have
side effects when the users do want to know if the operation was
actually successful.

Wrike Ticket: https://www.wrike.com/open.htm?id=1182861754