fix: sets default confirmation action to 'No'
Certain confirmations had the default action set to True. This would
mean that if someone hits the return key, then it will go ahead and
perform the action. It may lead to unwanted consequences and thus a
risky option to have in the CLI. This commit fixes that.

Wrike Ticket: https://www.wrike.com/open.htm?id=1299312443