fix(configtree): replace --project with --organization flag
The initial implementation of the configtree provided an option to
operate on project-scoped trees using the --project flag. However, after
considering the use cases, we have come to an understanding that it is
more common to operate on project-scoped trees than the
organization-scoped trees. Hence, this commit replaces the --project
with --organization flag. Specifying the flag means that the user wants
to operate on the organization-scoped config trees only.