fix(secret): corrects corresponding find functions (#301)
Due to a recent refactor in the v2 client library,
the list secrets function now returns the entire
secret objects instead of just reading the metadata
in the previous implementation. This changed object
altogther and the way other sections of code fetched
the secret name.

This commit addresses the issue by correcting the
find functions and other methods where secret name is
referred to.