feat(utils): extends show_selection helper
This commit updates the show_selection helper in selector.py to accept
additional arguments and extend its capabilities to show better
selection prompts to the user.

This change specifically adds two arguments, show_keys and
highlight_item.

If show_keys is set to true then for dict type inputs, the keys will be
presented to the user in the prompt.

If highlighter_item is passed in the args, then item matching the value
will be displayed in bold and italics.