fix(project): updates subnet when vpn is already enabled
The rio project features vpn command does not update the subnet if the
VPN is already enabled. The reason is a check that returns if the
current state is already the desired one.

This commit removes the check and allows an update to happen
irrespective and thus allowing the subnet range updates to work.

Wrike Ticket: https://www.wrike.com/open.htm?id=1418307735