✨ feat: implements update and delete for virtual devices (#326)
This commit improves the support for onboarding HWIL devices via device manifests. We implement the update_object and delete_object such that HWIL devices are re-onboarded or deleted via the rio apply and rio delete commands.

Take an example where the rapyuta.io device is present but the corresponding HWIL device pod was evicted, we will have to re-onboard it. On re-apply the manifest, the update_object handler will take care of running the onboarding script against on the said device. If in case the device was deleted as a part of some cleanup, then re-applying will ensure that the device is re-created and onboarded.

Similarly, rio delete device.yaml will ensure that both rapyuta.io and the HWIL device are cleaned up.

Wrike Ticket: https://www.wrike.com/open.htm?id=1425685033