fix(deployment): show errors in deployments list
The list deployment command isn't helpful to know at a glance the
actual status of the deployment. Users would have had to inspect
the deployments individually to know the errors and even then it
won't be helpful enough to derive the actual error. This commit
fixes the issue by parsing the error code and printing the error.

Wrike Ticket: https://www.wrike.com/open.htm?id=1449503225