fix(bootstrap): handle exceptions thrown by the cli
Exceptions aren't handled gracefully in the CLI. It's cumbersome to do
it at all the places. This fix wraps the entrypoint in a try-except
block and gracefully handles the exceptions.

You can set the DEBUG=true in order to continue seeing the tracebacks.

Wrike Ticket: https://www.wrike.com/open.htm?id=1382438801