fix(configtree): update sentinel key when tree is imported in etcd (#337)
When config files are updated to etcd via riocli, it doesn't update the sentinel key that the io-amr stack watches on to figure out config changes. Our ioconfig-syncer implementation does it automatically. In this commit, we ensure that the sentinel key is updated as a part of import.


Wrike Task: https://www.wrike.com/open.htm?id=1459385386