refactor(apply): remove apply summary and overview table
The apply command provides an overview of how much time it may take
and the actions that will be performed during apply. However, the
calculation was based on estimates and are incorrect. Similary, the
actions are also incorrect at many places and neither do we need it.

This commit removes the chunk of code that prints that info.

Wrike Ticket: https://www.wrike.com/open.htm?id=1183537395