refactor(apply): removes resolver cache
The purpose of the ResolverCache was to perform name to guid translation
from the dependencies defined in the manifests. However, the overall
implementation added an overhead of performing multiple API calls before
the actual call of creating or deleting a resource is made. This led to
a slowdown in the performance of the CLI when there are multiple
resources to create.

In this commit, we completely remove the resolver cache and refactor the
implementation to remove the overhead of making additional API calls.
We have also simplified to base Model class and now, the concrete
classes will be responsible for doing everything relevant for a
particular resource.

Wrike Ticket: https://www.wrike.com/open.htm?id=1450590338