fix(organization): fix organization select command (#339)
The organization select command checks whether the currently set
organization is already the selected one. It relies on the
organization_id in the config. However, there may be a case where the
config may not have the organization entry. In such a case, accessing
the key without .get() throws a key error. This commit rectifies the
issue.

Wrike Ticket: https://www.wrike.com/open.htm?id=1464951724