perf(delete): implements multi-threaded delete operation
The delete operation has been sequential, till date. For large number of
deployments, it takes quite a lot of time to perform a cleanup. The
problem was performing the deletes in the right order and the
TopologicalSorter does not provide a way to do that while maintaining
concurrency.

This commit implements the required logic to get the right delete order
while processing them concurrently.

Wrike Ticket: https://www.wrike.com/open.htm?id=1467788260