feat(configtree): supports overrides when importing config from files
This commit adds the support for config overrides while importing
configurations from files. Users can specify one or more override
files and they will be applied to the base configs in order. It is
expected that the overrides have the full key hierarchy such that
the command is able to match the right override with the base config.

Wrike Ticket: https://www.wrike.com/open.htm?id=1459374448