feat(deployment): implement restart alias for rio deployment update
The rio deployment update command essentially restarts the deployment
and serves no other purpose. Hence, it makes sense to rename the command
to restart. However, there may be scripts that are currently using the
update command and hence, in order to maintain compatibility, we
implement an alias in this commit and mark the update command as
deprecated.

Wrike Ticket: https://www.wrike.com/open.htm?id=1470267514