fix: corrects error message for batch delete ops
Most delete operations now run in batch mode. This also changes the way
the messages are printed. In the case when a single resource is deleted
and it fails, the commands print partial success. This is somewhat
misleading and this commit fixes that.

Wrike Ticket: https://www.wrike.com/open.htm?id=148581298