feat(scripts): adds install script for appimages
This commit introduces a script to download an install appimages. This
will provide convenience to the users of the CLI who want to download
and use the appimage.

It also adds it the CLI docs that are hosted via Github Pages. That
should allow access the files via https://cli.rapyuta.io/install.sh
and can be executed directly on a terminal providing an easier
installation option to the users.